initial semantic actions to ast (temporary)
