+++ src/stan/lang/grammars/bare_type_grammar.hpp
@@ -2,16 +2,15 @@
 #define STAN_LANG_GRAMMARS_BARE_TYPE_GRAMMAR_HPP
 
 #include <boost/spirit/include/qi.hpp>
-
 #include <stan/lang/ast.hpp>
 #include <stan/lang/grammars/whitespace_grammar.hpp>
-#include <stan/lang/grammars/statement_grammar.hpp>
 
 #include <string>
 #include <sstream>
 #include <vector>
 
 namespace stan {
+
   namespace lang {
 
     template <typename Iterator>
